refactor: rename Board domain to Ranking
