--- backend/src/main/kotlin/day/widdle/widdle/board/domain/Board.kt
@@ -1,55 +1,0 @@
-package day.widdle.widdle.board.domain
-
-import day.widdle.widdle.board.domain.vo.BoardId
-import day.widdle.widdle.board.domain.vo.Nickname
-import day.widdle.widdle.board.domain.vo.Statistics
-import jakarta.persistence.Column
-import jakarta.persistence.Embedded
-import jakarta.persistence.EmbeddedId
-import jakarta.persistence.Entity
-import jakarta.persistence.EntityListeners
-import jakarta.persistence.Table
-import org.springframework.data.annotation.CreatedDate
-import org.springframework.data.annotation.LastModifiedDate
-import org.springframework.data.jpa.domain.support.AuditingEntityListener
-
-@Entity
-@Table(name = "board")
-@EntityListeners(AuditingEntityListener::class)
-class Board(
-    @EmbeddedId
-    val id: BoardId = BoardId(),
-
-    @Embedded
-    var nickname: Nickname,
-
-    @Embedded
-    var statistics: Statistics,
-
-    @Column(name = "is_korean")
-    val isKorean: Boolean = true,
-
-    @CreatedDate
-    @Column(name = "created_at")
-    val createdAt: Long = 0L,
-
-    @LastModifiedDate
-    @Column(name = "modified_at")
-    var modifiedAt: Long = 0L
-) {
-
-    fun updateNickname(nickname: Nickname) {
-        this.nickname = nickname
-    }
-
-    fun updateStatistics(statistics: Statistics) {
-        this.statistics = Statistics(
-            totalStreak = statistics.totalStreak,
-            successRate = statistics.successRate,
-            currentStreak = statistics.currentStreak,
-            bestStreak = statistics.bestStreak,
-            todayPlaytime = statistics.todayPlaytime,
-            totalPlaytime = statistics.todayPlaytime + this.statistics.totalPlaytime,
-        )
-    }
-}
