--- backend/src/main/kotlin/day/widdle/widdle/board/domain/exception/BoardIdNotFoundException.kt
@@ -1,5 +1,0 @@
-package day.widdle.widdle.board.domain.exception
-
-import day.widdle.widdle.global.exception.WiddleException
-
-class BoardIdNotFoundException : WiddleException("id를 찾을 수 없습니다.")
