--- backend/src/main/kotlin/day/widdle/widdle/board/domain/projection/BoardWithRanking.kt
@@ -1,8 +1,0 @@
-package day.widdle.widdle.board.domain.projection
-
-import day.widdle.widdle.board.domain.Board
-
-interface BoardWithRanking {
-    val board: Board
-    val ranking: Long
-}
