--- backend/src/main/kotlin/day/widdle/widdle/board/service/dto/StatisticsDto.kt
@@ -1,21 +1,0 @@
-package day.widdle.widdle.board.service.dto
-
-import day.widdle.widdle.board.domain.Board
-
-data class StatisticsDto(
-    val id: String,
-    val rank: Int,
-    val nickname: String,
-    val playtime: Long
-) {
-
-    companion object {
-        fun of(board: Board, index: Int): StatisticsDto = StatisticsDto(
-            id = board.id.value,
-            rank = index,
-            nickname = board.nickname.value,
-            playtime = board.statistics.todayPlaytime
-        )
-    }
-
-}
