--- backend/src/main/kotlin/day/widdle/widdle/board/service/dto/StatisticsSaveDto.kt
@@ -1,10 +1,0 @@
-package day.widdle.widdle.board.service.dto
-
-import day.widdle.widdle.board.domain.vo.Nickname
-import day.widdle.widdle.board.domain.vo.Statistics
-
-data class StatisticsSaveDto(
-    val nickname: Nickname,
-    val statistics: Statistics,
-    val isKorean: Boolean
-)
