+++ backend/src/main/kotlin/day/widdle/widdle/ranking/controller/dto/StatisticsListResponse.kt
@@ -1,0 +1,8 @@
+package day.widdle.widdle.ranking.controller.dto
+
+import day.widdle.widdle.ranking.service.dto.StatisticsDto
+
+data class StatisticsListResponse(
+    val rankings: List<StatisticsDto>,
+    val me: StatisticsDto?
+)
