+++ backend/src/main/kotlin/day/widdle/widdle/ranking/domain/exception/DuplicateNicknameException.kt
@@ -1,0 +1,5 @@
+package day.widdle.widdle.ranking.domain.exception
+
+import day.widdle.widdle.global.exception.WiddleException
+
+class DuplicateNicknameException : WiddleException("닉네임은 중복될 수 없습니다.")
