+++ backend/src/main/kotlin/day/widdle/widdle/ranking/domain/exception/RankingIdNotFoundException.kt
@@ -1,0 +1,5 @@
+package day.widdle.widdle.ranking.domain.exception
+
+import day.widdle.widdle.global.exception.WiddleException
+
+class RankingIdNotFoundException : WiddleException("id를 찾을 수 없습니다.")
