+++ backend/src/main/kotlin/day/widdle/widdle/ranking/domain/projection/BoardWithRanking.kt
@@ -1,0 +1,8 @@
+package day.widdle.widdle.ranking.domain.projection
+
+import day.widdle.widdle.ranking.domain.Ranking
+
+interface BoardWithRanking {
+    val ranking: Ranking
+    val rank: Long
+}
