+++ backend/src/main/kotlin/day/widdle/widdle/ranking/service/dto/StatisticsSaveDto.kt
@@ -1,0 +1,10 @@
+package day.widdle.widdle.ranking.service.dto
+
+import day.widdle.widdle.ranking.domain.vo.Nickname
+import day.widdle.widdle.ranking.domain.vo.Statistics
+
+data class StatisticsSaveDto(
+    val nickname: Nickname,
+    val statistics: Statistics,
+    val isKorean: Boolean
+)
