+++ backend/src/main/kotlin/day/widdle/widdle/ranking/service/dto/StatisticsUpdateDto.kt
@@ -1,0 +1,9 @@
+package day.widdle.widdle.ranking.service.dto
+
+import day.widdle.widdle.ranking.domain.vo.RankingId
+import day.widdle.widdle.ranking.domain.vo.Statistics
+
+data class StatisticsUpdateDto(
+    val id: RankingId,
+    val statistics: Statistics
+)
